devdb_python/requirements.txt
databricks-sql-connector>=3.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
